fix(build): cast enum operands in two conditionals GCC -Werror rejects
GCC 15 with the first-party -Wall -Wextra -Werror set rejects two ternaries
the sync and its cleanup introduced:
- cypher.cpp: `is_with ? CYP_MAX_VARS : CBM_SZ_32` compares two different
  unnamed enums (-Werror=enum-compare);
- pass_complexity.cpp: `callee_cap ? callee_cap : CBM_SZ_64` mixes size_t with
  an enum (-Werror=extra).
Cast both to the destination type; values are unchanged.

Found by building every target with Homebrew gcc-15 after the Linux CI legs
failed. All first-party TUs now compile under GCC except four that only fail
inside the macOS SDK headers (cli, mem, platform, subprocess), which are
exercised by the Linux CI legs.

+++ src/cypher/cypher.cpp
@@ -1685,7 +1685,7 @@ static int parse_return_or_with(parser_t *p, cbm_return_clause_t **out, bool is_
      * and binding_t holds exactly CYP_MAX_VARS variables. A wider WITH used to
      * parse, then lose every alias past the 16th in with_add_vbinding_var and
      * answer with silently blank or missing columns. Refuse it here instead. */
-    if (r->count > (is_with ? CYP_MAX_VARS : CBM_SZ_32)) {
+    if (r->count > (is_with ? (int)CYP_MAX_VARS : (int)CBM_SZ_32)) {
         free_return_clause(r);
         return CBM_NOT_FOUND;
     }
